Pull out StakingTimelock component
